Move fixtures to pytest_fixtures
